[setup] Fonts: check if fonts need to be installed

+++ setup.sh
@@ -50,17 +50,21 @@ BREW="$LOCAL/bin/brew"
 echo "Homebrew packages installed."
 
 # Install fonts.
-echo "Installing fonts..."
-source_code_pro_url=$(curl -s "https://api.github.com/repos/adobe-fonts/source-code-pro/releases/latest" \
-                                            | fgrep zipball_url \
-                                            | tr -d ' ,"' \
-                                            | cut -d : -f 2-)
-curl -sL "$source_code_pro_url" -o source_code_pro.zip
-unzip -q source_code_pro.zip
-mkdir -p "$FONTS"
-find adobe-fonts-source-code-pro* -iname "*.ttf" \
-    | tr '\n' '\0' \
-    | xargs -0 -n 1 -I __file__ mv __file__ "$FONTS/"
-rm -rf adobe-fonts-source-code-pro*
-rm source_code_pro.zip
-echo "Fonts installed."
+if [[ ! -e "$FONTS/SourceCodePro-Medium.otf" ]]; then
+  echo "Installing fonts..."
+  source_code_pro_url=$(curl -s "https://api.github.com/repos/adobe-fonts/source-code-pro/releases/latest" \
+                                              | fgrep zipball_url \
+                                              | tr -d ' ,"' \
+                                              | cut -d : -f 2-)
+  curl -sL "$source_code_pro_url" -o source_code_pro.zip
+  unzip -q source_code_pro.zip
+  mkdir -p "$FONTS"
+  find adobe-fonts-source-code-pro* -iname "*.otf" \
+      | tr '\n' '\0' \
+      | xargs -0 -n 1 -I __file__ mv __file__ "$FONTS/"
+  rm -rf adobe-fonts-source-code-pro*
+  rm source_code_pro.zip
+  echo "Fonts installed."
+else
+  echo "Fonts are already installed."
+fi
